Revert "Streams"
