+++ directive/directive.go
@@ -12,7 +12,6 @@ import (
 // InputTypeRequest and others represent consts for Directives
 const (
 	InputTypeRequest = "request"
-	InputTypeStream  = "stream"
 )
 
 // Directive describes a set of functions and a set of handlers
@@ -53,7 +52,6 @@ type ScheduleEvery struct {
 // Input represents an input source
 type Input struct {
 	Type     string `yaml:"type" json:"type"`
-	Source   string `yaml:"source,omitempty" json:"source,omitempty"`
 	Method   string `yaml:"method" json:"method"`
 	Resource string `yaml:"resource" json:"resource"`
 }
@@ -187,8 +185,8 @@ func (d *Directive) Validate() error {
 	}
 
 	for _, h := range d.Handlers {
-		if h.Input.Type != InputTypeRequest && h.Input.Type != InputTypeStream {
-			problems.add(fmt.Errorf("handler for resource %s has invalid type, must be 'request' or 'stream'", h.Input.Resource))
+		if h.Input.Type == "" {
+			problems.add(fmt.Errorf("handler for resource %s missing type", h.Input.Resource))
 		}
 
 		if h.Input.Resource == "" {
