+++ go.mod
@@ -4,13 +4,11 @@ go 1.16
 
 require (
 	github.com/google/uuid v1.2.0
-	github.com/gorilla/websocket v1.4.2
 	github.com/pkg/errors v0.9.1
 	github.com/sethvargo/go-envconfig v0.3.2
 	github.com/spf13/cobra v1.1.3
 	github.com/suborbital/grav v0.3.3
-	github.com/suborbital/reactr v0.10.1-0.20210727154932-f316d5d35fd7
-	github.com/suborbital/subo v0.0.18
+	github.com/suborbital/reactr v0.10.1-0.20210703180801-dad2e23e943e
 	github.com/suborbital/vektor v0.4.1
 	golang.org/x/mod v0.4.2
 	gopkg.in/yaml.v2 v2.4.0
